requirements.txt
ultralytics
gradio
pillow
opencv-python-headless
numpy
